add view model in canvas
